Decouple length bubble to module
